refactor: Game 폴더 Packet에 관련 클래스 이동
