+++ RTWTest/Game/GamePacketFactoryTests.cs
@@ -1,6 +1,6 @@
 using Google.Protobuf;
 using RTW.NetworkDefinition.Proto.Packet;
-using RTWServer.Game;
+using RTWServer.Game.Packet;
 using RTWServer.Packet;
 
 namespace RTWTest.Game;
